Moves some tests from api/client/client/client_test to api/apiclient_test
