+++ api/apiclient_test.go
@@ -4,6 +4,7 @@
 package api_test
 
 import (
+	"bytes"
 	"context"
 	"crypto/tls"
 	"crypto/x509"
@@ -12,8 +13,10 @@ import (
 	"net"
 	"net/http"
 	"net/http/httptest"
+	"net/url"
 	"os"
 	"reflect"
+	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -21,6 +24,7 @@ import (
 	"github.com/juju/clock"
 	"github.com/juju/clock/testclock"
 	"github.com/juju/errors"
+	"github.com/juju/loggo"
 	"github.com/juju/names/v4"
 	proxyutils "github.com/juju/proxy"
 	"github.com/juju/testing"
@@ -28,6 +32,9 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/api"
+	"github.com/juju/juju/api/base"
+	apiclient "github.com/juju/juju/api/client/client"
+	"github.com/juju/juju/api/common"
 	apitesting "github.com/juju/juju/api/testing"
 	apiservertesting "github.com/juju/juju/apiserver/testing"
 	"github.com/juju/juju/controller"
@@ -1203,6 +1210,177 @@ func (s *apiclientSuite) TestLoginIncompatibleClient(c *gc.C) {
 		"juju client with major version %d used with a controller having major version %d not supported\\n.*",
 		jujuversion.Current.Major, 99,
 	))
+}
+
+func (s *apiclientSuite) TestConnectStreamRequiresSlashPathPrefix(c *gc.C) {
+	reader, err := s.APIState.ConnectStream("foo", nil)
+	c.Assert(err, gc.ErrorMatches, `cannot make API path from non-slash-prefixed path "foo"`)
+	c.Assert(reader, gc.Equals, nil)
+}
+
+func (s *apiclientSuite) TestConnectStreamErrorBadConnection(c *gc.C) {
+	s.PatchValue(&api.WebsocketDial, func(_ api.WebsocketDialer, _ string, _ http.Header) (base.Stream, error) {
+		return nil, fmt.Errorf("bad connection")
+	})
+	reader, err := s.APIState.ConnectStream("/", nil)
+	c.Assert(err, gc.ErrorMatches, "bad connection")
+	c.Assert(reader, gc.IsNil)
+}
+
+func (s *apiclientSuite) TestConnectStreamErrorNoData(c *gc.C) {
+	s.PatchValue(&api.WebsocketDial, func(_ api.WebsocketDialer, _ string, _ http.Header) (base.Stream, error) {
+		return api.NewFakeStreamReader(&bytes.Buffer{}), nil
+	})
+	reader, err := s.APIState.ConnectStream("/", nil)
+	c.Assert(err, gc.ErrorMatches, "unable to read initial response: EOF")
+	c.Assert(reader, gc.IsNil)
+}
+
+func (s *apiclientSuite) TestConnectStreamErrorBadData(c *gc.C) {
+	s.PatchValue(&api.WebsocketDial, func(_ api.WebsocketDialer, _ string, _ http.Header) (base.Stream, error) {
+		return api.NewFakeStreamReader(strings.NewReader("junk\n")), nil
+	})
+	reader, err := s.APIState.ConnectStream("/", nil)
+	c.Assert(err, gc.ErrorMatches, "unable to unmarshal initial response: .*")
+	c.Assert(reader, gc.IsNil)
+}
+
+func (s *apiclientSuite) TestConnectStreamErrorReadError(c *gc.C) {
+	s.PatchValue(&api.WebsocketDial, func(_ api.WebsocketDialer, _ string, _ http.Header) (base.Stream, error) {
+		err := fmt.Errorf("bad read")
+		return api.NewFakeStreamReader(&badReader{err}), nil
+	})
+	reader, err := s.APIState.ConnectStream("/", nil)
+	c.Assert(err, gc.ErrorMatches, "unable to read initial response: bad read")
+	c.Assert(reader, gc.IsNil)
+}
+
+// badReader raises err when Read is called.
+type badReader struct {
+	err error
+}
+
+func (r *badReader) Read(p []byte) (n int, err error) {
+	return 0, r.err
+}
+
+func (s *apiclientSuite) TestConnectControllerStreamRejectsRelativePaths(c *gc.C) {
+	reader, err := s.APIState.ConnectControllerStream("foo", nil, nil)
+	c.Assert(err, gc.ErrorMatches, `path "foo" is not absolute`)
+	c.Assert(reader, gc.IsNil)
+}
+
+func (s *apiclientSuite) TestConnectControllerStreamRejectsModelPaths(c *gc.C) {
+	reader, err := s.APIState.ConnectControllerStream("/model/foo", nil, nil)
+	c.Assert(err, gc.ErrorMatches, `path "/model/foo" is model-specific`)
+	c.Assert(reader, gc.IsNil)
+}
+
+func (s *apiclientSuite) TestConnectControllerStreamAppliesHeaders(c *gc.C) {
+	catcher := api.UrlCatcher{}
+	headers := http.Header{}
+	headers.Add("thomas", "cromwell")
+	headers.Add("anne", "boleyn")
+	s.PatchValue(&api.WebsocketDial, catcher.RecordLocation)
+
+	_, err := s.APIState.ConnectControllerStream("/something", nil, headers)
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(catcher.Headers().Get("thomas"), gc.Equals, "cromwell")
+	c.Assert(catcher.Headers().Get("anne"), gc.Equals, "boleyn")
+}
+
+func (s *apiclientSuite) TestWatchDebugLogParamsEncoded(c *gc.C) {
+	catcher := api.UrlCatcher{}
+	s.PatchValue(&api.WebsocketDial, catcher.RecordLocation)
+
+	params := common.DebugLogParams{
+		IncludeEntity: []string{"a", "b"},
+		IncludeModule: []string{"c", "d"},
+		IncludeLabel:  []string{"e", "f"},
+		ExcludeEntity: []string{"g", "h"},
+		ExcludeModule: []string{"i", "j"},
+		ExcludeLabel:  []string{"k", "l"},
+		Limit:         100,
+		Backlog:       200,
+		Level:         loggo.ERROR,
+		Replay:        true,
+		NoTail:        true,
+		StartTime:     time.Date(2016, 11, 30, 11, 48, 0, 100, time.UTC),
+	}
+
+	urlValues := url.Values{
+		"includeEntity": params.IncludeEntity,
+		"includeModule": params.IncludeModule,
+		"includeLabel":  params.IncludeLabel,
+		"excludeEntity": params.ExcludeEntity,
+		"excludeModule": params.ExcludeModule,
+		"excludeLabel":  params.ExcludeLabel,
+		"maxLines":      {"100"},
+		"backlog":       {"200"},
+		"level":         {"ERROR"},
+		"replay":        {"true"},
+		"noTail":        {"true"},
+		"startTime":     {"2016-11-30T11:48:00.0000001Z"},
+	}
+
+	client := apiclient.NewClient(s.APIState)
+	_, err := client.WatchDebugLog(params)
+	c.Assert(err, jc.ErrorIsNil)
+
+	connectURL, err := url.Parse(catcher.Location())
+	c.Assert(err, jc.ErrorIsNil)
+
+	values := connectURL.Query()
+	c.Assert(values, jc.DeepEquals, urlValues)
+}
+
+func (s *apiclientSuite) TestWatchDebugLogConnected(c *gc.C) {
+	cl := apiclient.NewClient(s.APIState)
+	// Use the no tail option so we don't try to start a tailing cursor
+	// on the oplog when there is no oplog configured in mongo as the tests
+	// don't set up mongo in replicaset mode.
+	messages, err := cl.WatchDebugLog(common.DebugLogParams{NoTail: true})
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(messages, gc.NotNil)
+}
+
+func (s *apiclientSuite) TestConnectStreamAtUUIDPath(c *gc.C) {
+	catcher := api.UrlCatcher{}
+	s.PatchValue(&api.WebsocketDial, catcher.RecordLocation)
+	model, err := s.State.Model()
+	c.Assert(err, jc.ErrorIsNil)
+	info := s.APIInfo(c)
+	info.ModelTag = model.ModelTag()
+	apistate, err := api.Open(info, api.DialOpts{})
+	c.Assert(err, jc.ErrorIsNil)
+	defer apistate.Close()
+	_, err = apistate.ConnectStream("/path", nil)
+	c.Assert(err, jc.ErrorIsNil)
+	connectURL, err := url.Parse(catcher.Location())
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(connectURL.Path, gc.Matches, fmt.Sprintf("/model/%s/path", model.UUID()))
+}
+
+func (s *apiclientSuite) TestOpenUsesModelUUIDPaths(c *gc.C) {
+	info := s.APIInfo(c)
+
+	// Passing in the correct model UUID should work
+	model, err := s.State.Model()
+	c.Assert(err, jc.ErrorIsNil)
+	info.ModelTag = model.ModelTag()
+	apistate, err := api.Open(info, api.DialOpts{})
+	c.Assert(err, jc.ErrorIsNil)
+	apistate.Close()
+
+	// Passing in an unknown model UUID should fail with a known error
+	info.ModelTag = names.NewModelTag("1eaf1e55-70ad-face-b007-70ad57001999")
+	apistate, err = api.Open(info, api.DialOpts{})
+	c.Assert(errors.Cause(err), gc.DeepEquals, &rpc.RequestError{
+		Message: `unknown model: "1eaf1e55-70ad-face-b007-70ad57001999"`,
+		Code:    "model not found",
+	})
+	c.Check(err, jc.Satisfies, params.IsCodeModelNotFound)
+	c.Assert(apistate, gc.IsNil)
 }
 
 type clientDNSNameSuite struct {
